[mac_fonts] Separate each API
