+++ find_system_fonts_filename/mac/types.py
@@ -1,0 +1,5 @@
+from ctypes import c_long, c_uint32, c_void_p
+
+CFIndex = c_long
+CFStringEncoding = c_uint32
+CFNumberRef = c_void_p
